Move siteSection type and type guard into model repo and fix linkResolver to account for untransformed data
